# hugolib: Avoid recloning of shortcode templates